+++ src/views/Consumer/Notifications.js
@@ -1,11 +1,11 @@
 import * as React from 'react';
 import {Button, View, Text} from 'react-native';
-import {ConsumerHeader} from './Components/ConsumerHeader';
+import { AppHeader } from '../Common/AppHeader';
 
 export const Notifications = ({navigation, title}) => {
   return (
     <View>
-      <ConsumerHeader navigation={navigation} title="Notifications" />
+      <AppHeader navigation={navigation} title="Notifications" />
       <Text>Notificatins</Text>
     </View>
   );
